A lot of things have changed.
The threadedserver is not used anymore since a server is not needed.
Instead of that an asynchronous TCP socket is enough implemented in async.py

Connect was just another attempt for asynchronous TCP sockets. I just upload it
now for completeness. In the next commit is going to be deleted.

These changes are propagated also to the proxy file.

In this version the main component myof_01 is not connected with the rest of the
modules but has the desired functionality by itself. The modifications allow
to make the central controller believe that the switch has connected to it.